requirements.txt
# tkinter (bundled with Python)
bcrypt
python-dotenv
psycopg2-binary
sqlite-utils
pytest
pytest-cov
